Switch out AttributeReportBuilder
